Create the class TieBraker to only analyzed the high pokerhand and refactor the other class to almost follow the rules of every method and class
