+++ src/main/java/jalau/at18/katas/pokerhands/daniela/Winning.java
@@ -1,6 +1,6 @@
 package jalau.at18.katas.pokerhands.daniela;
 
-public class Winning {
+public class Winning extends TieBraker {
 
     public Winning() {
     }
@@ -16,22 +16,9 @@ public class Winning {
         GameAnalizer winner1 = new GameAnalizer(black);
         GameAnalizer winner2 = new GameAnalizer(white);
         if (winner2.getscore() == winner1.getscore()) {
-            // untie(white, black, whitecard, blackcard);
             return untie(white, black, whitecard, blackcard);
         }
         return "There is a winner already";
-    }
-
-    public int highCardWhite(PokerHand handwhite, CardValue cardwhite) {
-        Cardssorted counter = new Cardssorted(handwhite);
-        CardstoString count = counter.getCount();
-        return count.highCard(cardwhite);
-    }
-
-    public int highCardBlack(PokerHand handBlack, CardValue cardBlack) {
-        Cardssorted counter = new Cardssorted(handBlack);
-        CardstoString count = counter.getCount();
-        return count.highCard(cardBlack);
     }
 
     public String untie(PokerHand white, PokerHand black, CardValue whitecard, CardValue blackcard) {
